Working on the music generator
1. Refactored and expanded ProcMu data structures to prepare implementation of various generative algorithms.
2. Started implementing algorithms for music generation. Random, Perlin, LSystem --> see documentation.
3. Implemented GUIs for the ProcMu features. Developing the UI along with the new features, to be able to configure the new parameters.

Still some technical work to be done here but getting close to pleasant sounding music. Once that is dealt with, I will implement the gameplay. On this note, considering to not interpolate musical styles between music zones but concentrate on rhythmic synchronization and integration of spatialized objects with the music.