Started work on the (new) interpolator and cleaned up folder structure
Since basic music generation is working, the next step is to introduce change. For now, this change should be dependent on the player position in the game world. Music zones are areas that hold their own generative music configuration. The interpolator is responsible for creating a generative music configuration based on the configurations of all MusicZones around the player. For example, the parameters that define a rhythmic pattern will be calculated based on the weighted average of all given configurations. The weight is defined by the distance between a zone and the player. The closer the zone, the higher its configuration impact.
I took some code from ProcMu's interpolator but there is still plenty to do in terms of interpolating each variable in the configurations. Some values will not be interpolated. Where it makes sense, I will just take the value of the closest music zone. For example, when choosing between discrete functions such as arpeggiator direction.