Removing Unity project
Audio output with AudioHelm did not work in builds which is a dealbreaker. Hence it's back to Unreal where MetaSound seems to have all I need. I put together a basic setup, which after two days of work is already ahead of the Unity approach. Builds work as well. Therefore, I'm removing the Unity project. The next commit will push the Unreal project.
Also, renaming the journal directory to docs for Github pages support in an attempt to make the journal experience nicer.