Uploading the unreal project
There is actually a lot of stuff that happened here. This is basically what I had already made in Unity, plus working sound objects but minus music zones and interpolation. Also there's now properly functioning procedural sound, a working build, and of course everything audio now using MetaSounds. There is a new island with different areas: a beach, a cove, and a hill, as well as paths in between these areas. Each area holds a signature sound object. The next step is to tune the existing beach, cove, and hill sound objects to sound interesting, fitting, and pleasant. The idea here is to have two parts: 1. a part of the sound that aims to represent nature (e.g., wind, waves, splashes...) 2. a musical part that is harmonically linked to the overall music construct. In addition to these sound objects, there is non-diegetic-like music that is supposed to tie everything together. This main music still needs to be set up, including chord, rhythm, and melody generation.