Dressed up the island, iterating on the gameplay
Added textures and 3d assets to give the island a visual style, as well as to define the paths the player can walk.
I'm also working to make the gameplay more interesting. The goal object is now placed on the ground, at a random location around the designated spawn area. Next, I want to implemented the player having to press a key to pick up the object. The object should be randomized. It could be a stone, or something else that fits the environment.
Then, I want to make the music system reflect the player's relation to the object (e.g., distance to the object, area object is located in).