World design, more musical elements, enhanced gameplay
I've populated the world with assets, most greyboxing is gone now. There are more sounds in the music and I've added some first connections between player behavior and the system: The music now changes when the player moves closer to the center of the cove area. I'm going to implement similar approaches in the other areas as well. The player can now only access the full map by collecting objects and thus removing barriers. The collectibles are now more subtle and hidden in plain sight, spawning randomly around designated areas. I think I'm set with the gameplay for now, so I'm going to focus all on the music over the next sessions.