Footsteps multi surface support but broken, easier tutorial
Fixed footsteps with added support for rock and sand surfaces and then broke it again. Also moved footsteps sound to its own object on the player character and out of the Soundgarden metasound for now. This should make troubleshooting easier. I'm also getting an idea for expanding footsteps to other parts of the body interacting with the environment: shoulders or legs rubbing against close surfaces.
Tutorial enhancements: Added two more collectables in the swamp area. This should reduce the likelihood of the player losing sight of where to go and find the path uphill better.