Writing and breaking lighting
Added some more information to the documentation. Unfortunately, the lighting completely broke upon reopening the project today and I really don't know why. Couldn't find anything relevant online, which leads me to think it could be a bug in the engine itself. Doing this backup commit to check if the problem persists when opening the project in the UE 5.4 preview. If it works, the new engine version should also bring a performance boost, since the developers claim to have made optimizations to Lumen. Fingers crossed.