Fixing things and journalings
I'm slowly getting the lighting back to how it was before. Nights are dark again, however, the sky is still brighter than before. It feels too bright when the sun isn't even on the horizon. Also added another journal entry.
Switching to 5.4 made it impossible to downgrade the map. I could pull an older version from before the switch, however, I've already made so many little changes that I will stick with 5.4 for now. Since it's scheduled to come out of preview at the end of this month, I'm hopeful the other new small issues will be fixed then.