Back to 5.3
Okay, 5.4 was causing trouble. Found a workaround for downgrading the project. We're back to 5.32 now. Everything seems to be working again but I should watch out for any issues I might have missed. Tamed the auto exposure issue. One of the causes was that the global directional light intensity was turned up way too high because of some bug...however, I still need to find a way to tone down the brightness of the sky at night.