Build troubles continued
Upgrading to 5.4 and then downgrading to 5.3 broke builds. Couldn't find a solution to this issue so doing this backup commit before going back to 5.4 again. I'd rather deal with a temporary performance regression than not being able to push new builds. 5.4 should come out of preview by the end of the month. May it'll fix the performance regression...