Moving music over to the game environment
I've begun to integrate the output of the Soundgarden music system into the game environment using speaker objects. Used a horn and bell asset from Quixel. I've turned the volume in the main MetaSound patch to zero. Might remove the mixer altogether.
Created two speaker variations to carry the sounds of voice 0 and 1. Scattered speakers around the swamp and parts of the hill. Some speakers don't seem to be working though. It's the ones which seem to be out of attenuation range at the start of the game. Need to check if there's some auto disabling going on...