Added walking/running, outlook adjustments, various audio improvements
The player now walks more slowly by default because there is now a running feature, which makes them move around the island faster than before. 
Changed up the look of the outlook, opening up the top and making it look more like rocks balanced on top of each other.
Adjusted attenuation and stereo spread on speaker objects. Added bass sound to the environment as two audio sources. Should play more with the idea that there are resonators underneath the island, which is where the sources are located. Also even more documenting.