Cleaned up sound patches, finished documentation
Patches look a bit nicer now. All the basic documentation is done, though I plan to add more elements in the future as there will likely be new sound modules or modifications of the existing ones. I also noticed that the parrot sound has some strange behavior when sun intensity is 0, leading to lots of cascades in the night. Must fix.