Fixed time triggers and parrots
Time triggers now hide and show properly. Also added pause, play, fast forward icons to show their impact on time. Parrots were calling at 0 intensity, which is fixed now. The reason for this bug was that the number of triggers would be set to 0 at 0 intensity, which means infinite repeats for the trigger module. Now, the minimum trigger amount is 1, but triggers will only be forwarded when intensity > 1.