Euclidean rhythms and new broken main menu
Added two Euclidean rhythm modules, one that works with float arrays, another that is a real-time version which will return a hit or not based on a given step. Replaced the Perlin noise based rhythm generators in the percussion module. The number of hits in each of the four percussive tracks is modulated by the intensity parameter. However, the game is currently broken because I also added a main menu. Still need to figure out how to start the game. Once that is fixed, the new version will go up on itch.