Fixed and added main menu
The main menu is open on game start and closes when clicking Play. There's also a quit option and a placeholder for a future settings menu. The default rendering resolution is currently a bit too low for capable hardware, and may be too high for low to mid tier devices, so this is one of the options that players should be able to change in the settings. After this little bit of polish, I should next add another event that's unlocked when collecting rocks. Also, the footstep sounds seem to be coming more from the left, which I should look into.