Added rhythm grid
As a first gameplay element to interact with Rhythmer, I have added the rhythm grid. The rhythm grid is a system that consists of a grid where each cell can be inactive, transparent, or active. The player can stand on an active cell. The idea is to have a defined trace of cells triggered successively, step-wise by a rhythm input. Currently testing with a 4x4 grid. Next up is to expand the grid and make a basic music setup whose rhythm controls the cell trace.