Rhythm grid enhancement
Adjusted the rhythm grid to preview the next usable cells. On the downbeat, the next cell is activated for the player to stand on. On subsequent beats, the upcoming cells will be previewed for the player to orient themselves where to step on the next downbeat. Currently this is set to work for a 4 note cycle but can be changed to work with other measures and rhythmic patterns as well.