Integrated rhythm grid into musicballs scene
Added concrete walls and steps leading up to and away from the rhythm grid. The concrete walls are meant to hide parts of the grid to make the overall path less obvious, as well as to prevent the player from jumping across the grid and possibly finding shortcuts. Once I have found a good configuration for the walls, I will integrate the rhythm grid into a second island in the main Soundgarden scene. With the second island, I aim for a less natural look, going for a more brutalist approach - concrete everywhere! The minimalist concrete look may help to emphasize the upcoming audiovisual elements that are rhythmically synchronized with the central Rhythmer system.