Added Wavefolder
Implemented a MetaSound patch for wavefolding, which allows to change the timbre of a sound by just changing one parameter, i.e., the fold amount. This will be used for the rhythm grid puzzle, to indicate which platforms in the grid can be used. I'm also planning to let platforms light up briefly in the rhythm, so that the player will be able to associate a given timbre with the platform lighting up.