Improved performance, fixed sound sources, minor adjustments
Performance is now strongly improved, mainly due to adjusting the global illumination settings to medium. Higher values caused a Lumen update every few frames which led to a massive framerate drop. Changing this setting required adjusting existing light sources and adding new ones to account for the overall darker scene. There are now some new setting-related nodes in the game controller blueprint which I can later use as a base for implementing user adjustable graphics settings.
I'm now using source buses for the all audio sources that use a separate MetaSound. This has a smaller overhead and delay compared to the previous solution, lowering CPU usage and further improving overall performance. The BusReader blueprints and MetaSounds are not required anymore. I have removed function calls here for now and will remove these objects altogether in the future. On that note, I should review which assets can be removed from the project to keep everything nice and tidy.
Based on the playtesting comments I got, I want to add more feedback on collecting stones. As a first step, the instruction painting in the swamp is now animated, lighting up when a stone is collected. Next, I want picking up a stone to be reflected in the music.