More wind path gameplay optimizations
I have modified the placement of the audio sources along the wind path's spline. Now, the front and rear sources are separated into two objects that are placed dynamically around the closest position to the player on the spline. Before, the sources were placed on a rig, essentially a rectangle, with the sound sources placed each corner. This rig would move along the spline. This had the disadvantage that sources would need a lot more space to be able to move around a curve without clipping through the surround walls of the tunnel. Furthermore, if the player was positioned between the front/rear sources and approaching a turn, the rig would only start turning when the player is at the very bend of the turn, making the required change of direction less clear. With the improved setup, front/rear sources can be placed further away from the player, while being able to turn around corners. Because of that, changes of direction can be heard more clearly.