Implemented saving
I added a feature that saves the game state. As of now, player transform, score, time of day, and unlocked areas are stored whenever the player collects a new stone after finishing the tutorial. In the future, we might need more save points once the other two islands are implemented. When the game is restarted, the save game is supposed to be loaded automatically. While the logic for everything is in, the actual saving and loading does not seem to work yet. I'm sure I'll have that fixed in the next commit. Once saving works, I will start building the sound-led island in the main scene and add the existing rhythm and wind path game mechanics.