Island 2 and save fix
I started implementing island 2 which will feature the sound-led interactions. I created the basic landscape and have begun dressing it up with rocks. I also created an outdoor labyrinth and a cave which will be the area that will feature the wind path gameplay. Everything is ready for that - I just need to add the wind path blueprint objects and make a spline for the outdoor labyrinth.
I also fixed saving. While that is working now, there are still some issues in restoring the game state after loading. Player positioning works but the day-night cycle and stone collection are broken. I also want to add an option to reset the save game and I want to place that in the settings menu, which I still need to create. However, once that is done I will also have everything ready to implement graphics quality settings.
I have also experimented with some new dynamic/reactive to music visual elements, specifically tree roots that are meant to glow and animate when it's dark. I'm not sure yet whether I will actually add this, there's still some more experimentation to do.