Finished wind path gameplay and hill
Added the wind path audio objects for the cave spline and created another spline for the outdoor labyrinth wind path section, which means that the wind path gameplay is up and running. Also finished dressing up the cave path hill with rocks. However, I got a bit sloppy so I might need to fix the placement of the rock slabs here and there.