New teleport, wind path tweak, and some fixes
Replaced the grey box path with teleporters. Players are now instantaneously transported between islands. This is only temporary, until the addition of the third gameplay area that will be placed in between islands 1 and 2. Tweaked the wind path so that it fades in the wind sound for the closest path end. This means that now when the player enters a wind path, they will first hear the wind sound that guides them to the other end. As they walk further down the path, the wind sound for the end they came from will gradually fade in.
Also fixed some colliders to avoid players getting stuck.
Attempted to fix the lighting being broken when the game is started without a saved game. Lighting works after reloading the level but the question is, how have it work without this workaround? Loading a dummy preloader level did not help...