Fixed broken lighting at game start
It's fixed. I added a BeginPlay event to the SunSky object that triggers a recalculation of lighting after a short delay. While this works, this causes a tiny hiccup and flash at the game start. Since audio has been coming in too rapidly anyway, I should eventually add a fade in for both the visuals and the sound to mask that.
Side note: I noticed that when modifying Unreal objects in the Plugins folder, this is not tracked on Git. I have to make sure to transfer all modified data from the Plugin folder to the game's content directory.