Added pickup interaction, fixed teleport visuals, added rhythm speakers
I have finished the animated sound speaker for rhythm playback and placed a few sources in the scene. I am still not quite happy with the visuals and consider scanning an object I found irl that would fit the desired look of the rhythm speaker, using photogrammetry.
I have implemented the basic system for context-sensitive interactions. When the player nears an interactable object, a symbol indicating the type of interaction is shown. Currently, this is only supported for picking up rocks. This also means that rocks are no longer picked up by walking over them. It also  would be nice to have a fitting sound that can be heard on pickup. I am also considering making the appearance of the rocks more consistent and thus make them easier to recognize. This would also simplify the implementation of a cairn mechanic, where collected rocks would be visualized in a pile.
I have also fixed the teleport visuals, making them visible in any lighting conditions. They also do not disappear anymore without reappearing again when the player moves too far away.