OddOneOut basics done, island 2 modeling, project cleanup
The odd one out game mechanic is basically done. I changed it from having to walk through one of three gates to interacting with one of three listener objects that each play a sound when the player is close. While randomization does seem to work according to the output of the controller blueprint, it is not reflected in the actual distribution of sounds for each listener. Probably just a small oversight somewhere I need to fix. The current implementation only supports disconnected setups of three listeners and one controller, respectively. I could either add more such setups, or just add more listeners to other areas that the controller dynamically switches between. I also still need to make barriers disappear when interacting with the correct listener. For that, and for the disappearing barriers in the tutorial section, i would like to add a disappearing shader effect which I have already built. Just need to implement a blueprint that controls the effect. Making the barriers slowly disappear with bright patterns could also help directing more attention to newly opened up areas on island 1.
I have also continued modeling the remaining areas of island 2, mainly the odd one out section. Now, this section leads to the landmass on the other side of the ravine, which is intended for the rhythm grid gameplay.
I have further optimized the project's folder structure. However, I noticed that moving and renaming folders leaves the original locations in the project, with the folders still holding files when checked in the explorer. This also throws reference error messages when attempting to delete them. I need to try removing them anyway, ideally right after a commit, to see if this causes any issues.
Next, I'm going to finish the odd one out mechanic, add a little jungle to the area, with fitting but interfering sounds that are intended to make it harder to discern the odd sound source.