OddOneOut working, animated barriers
The odd one out game mechanic is now fully implemented, supporting multiple controllers and controlling barriers. Now, I just need to add the remaining four controller sections in the area. Next, I will need to make everything look more like a jungle, to justify the (to be added) dense soundscape meant to increase difficulty in listening to the sounds.
On a side note, there are also new barrier objects that have a dissolve animation, which not only work in the OOO area but are also meant to replace the barriers in the tutorial. Their dissolution effect visually draws attention, which should help point out newly opened areas in the tutorial. This might especially help with seeing the opened path leading to the outer rim of the first island when standing atop the hill.