Finished OddOneOut, day-night system overhaul, various tweaks
I finished the OOO area. It didn't become quite the jungle in the end but grass and trees certainly make engulfed enough to justify a lush soundscape. Currently, only the sounds of the ocean and the rustling of leaves interfere with the listening gameplay but I plan to add more here. I have also worked on the sounds of this gameplay, with dynamic difficulty and randomization, making for a slightly different experience every time the puzzle is reset. I still find it a bit too easy, however, I'm waiting for feedback from players to see if that really is the case.
I have also added a night sky with a moon that brightens the nights. Therefore, I have also removed the light orb on island 1, since nights are now bright enough to navigate the environment without additional light. Since I have also tweaked the brightness of the directional light sources, I also had to make some adjustments for the auto exposure. Right now, the eye adaptation is basically off, since I'm still struggling to find the right settings that don't turn everything completely dark when the player looks at a dark area. To solve the issue of some areas being to dark, I have decreased the shadow contrast in the local exposure settings of the global post processing volume. I'm not sure if this solution is ideal, since it does make shadows less pronounced but perhaps it's just a matter of tweaking settings a bit more.
There have also been various other changes, improvements, additions, etc. Since I'm now also tracking tasks in a GitHub project board where I also document experiences and outcomes, I'm hereby referring to that for more details.
One more thing to mention is that performance should now be generally better, especially for the high end setting, due to decreasing cubemap resolution from 2048 to 128 in the SkyLight of the SunSky object.