UE 5.6 upgrade, started new journal entry, investigating sound changes
Soundgarden is now running on the latest version of Unreal Engine. I was hoping for some additional performance improvements here, but I'm not sure if that is actually the case. Still, it seems that things are still running just as well as before.
One thing I noticed, however, is that the melodic ocean synth (heard around island 1) sounds slightly different now. I can hear some phasing and more higher frequencies. I'll need to check if the sound filtering is still working, and if perhaps the sound source got doubled due to some changes I had made for the ocean sound speakers around island 2 later on. There's also a chance this is coming from updates to the MetaSound plugin itself - perhaps some features changed in the low pass filter node. Nevertheless, I wouldn't say the new sound is bad - I actually don't like it less than how it was before and perhaps it's even a bit more interesting due to the added phasing. Still, I would like to find out what's going on.
On that note, a reminder that I should start looking into implementing reverb (and finally learn how to do that in UE)...